clk: qcom: Hook up gpu_gx gdsc as a parent of gpu_gdsc
the GPU device on msm8996 requires 2 gdscs to be powered on.
With powerdomains we have a limitation wherein only one
powerdomain can be associated with a given device in DT.
So hook the 2 gdscs together so both are powered on/off when
the gpu driver needs them to.

[Amit] We use this hack until genpd allows multiple powerdomains to be
associated with a device.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>