drm/panel: Add driver for LG Display XXXXXXXXX FHD CMD panel
This is a driver for the LG Display XXXXXXXXX IPS-LCD DSI
command-mode panel, found in some Sony Xperia X and X Performance
smartphones.
The panel has a 1920x1080 (1080p "Full HD") resolution and uses
24 bit RGB per pixel.

This panel uses an in-cell regulator, which is activated with
two power supplies, vddio and avdd, and the power path to
respectively panel and integrated touch controller is activated
through two "pvddio" and "tvddio" GPIOs.
It also requires a specific reset sequence for both the LCD part
and the touch controller power part, through specific panel and
touch-controller reset "preset" and "treset" GPIOs.

That's the reason why this panel wasn't inserted in simple-panel
but, instead, a specific driver is required.