REFACTOR_ME: regulator: Add cpr2-gfx and mem-acc regulators for 8956 GPU access
Add these two regulators from downstream to fill a hole in this early
porting of the Xperia MSM8956 family of devices.

The drivers have been fast-forwarded to 5.3-rc2 and the CPR2 will now
accept a (genpd) power domain for the VDD_MX (now pd_mx), since that's
the only and PROPER way to do this.

Actually, this is both a good and a bad idea: having CPR2 and mem-acc
is a good idea, but having them LIKE THIS is a reeeeally bad idea.
The drivers should get partially refactored to register themselves
as power domains, NOT AS REGULATORS!!!!!!!

THIS COMMIT IS HERE FOR DEMO PURPOSES ONLY - THIS IS A TODO :)