drm/msm: Try to use DMA attached IOMMU domain before allocating new
The DMA API (at least on ARM/ARM64) tries to attach devices to its
own domains at of_xlate stage: these IOMMU domains would be left
unused, but the main problem here is that some IOMMUs for at least
some Qualcomm SoCs have static IOMMU contexts with only one
allocatable domain.
If the DMA API allocates one, then we're left with zero (so we are
left without an IOMMU context to use here); for this reason, check
if we can consume an IOMMU domain that was already pre-attached at
of_xlate() stage and, only if not, allocate a new one like before.

This solves the unability to get a working IOMMU domain on at least
SoCs using the qcom smmu-v2 and qcom mmu-500.

Signed-off-by: Angelo G. Del Regno <kholk11@gmail.com>