phy: qualcomm: Only disable regulators when cable was not plugged in.
This keeps track of whether the cable was plugged in when the phy is
enabled. The extcon will allways trigger before power on/off, thus
leaving cable_connected in the wrong state when the phy is powering off
due to an unplug event, in turn creating an unbalanced disable warning.

Now the regulators will only be disabled when they were enabled
initially.

TODO: Might want to move a boolean to the respective
qcom_snps_hsphy_{enable,disable}_regulators methods?