drm/panel: Add support for Xperia X Compact JDI 6 panel driver
This patch adds support for the JDI "6" panel found on the Sony
Xperia X Compact device.
This JDI panel is under a "Hybrid In-Cell" Renesas/Synaptics
R63452 Driver IC and has a resolution of 720x1080x24bit RGB
and some unknown amount of RAM (as that information is nowhere
to be found).